+++ lib/shellfie/parser_validation.rb
@@ -16,7 +16,8 @@ module Shellfie
     FONT_KEYS = %i[family size line_height fallback_family italic_family emoji_family].freeze
     ANIMATION_KEYS = %i[
       typing_speed command_delay cursor_blink loop typing_jitter seed typing_chunk_size output_delay final_delay max_frames
-      dither palette scroll_easing
+      dither palette gif_colors gif_optimize webp_lossless webp_quality webp_method webp_near_lossless
+      apng_prediction loop_count scroll_easing
       framerate playback_speed
     ].freeze
     CURSOR_KEYS = %i[style color].freeze